Create Checksum Method
Whatever data we want to send, the data is added up and the sum is sent along with the data (concatenated).
Example data: (7, 11, 12, 0, 6). The sum is 36.
The total data that will be sent is (7, 11, 12, 0, 6, 36).
The receiver takes the data, computes the sum again, compares both the sum values and if they match, it extracts the data. Otherwise, it discards.

Improvised method:
The sum is not sent along with data, instead, its checksum is sent i.e., its compliment (negation of sum) is sent. (7, 11, 12, 0, 6, -36). At the receiver side, it adds up all the values sent (including the checksum) and sees if the sum is zero. If there are no errors while transmission, the sum will be zero. 

We take the original data as 4-bit numbers i.e., a 4-bit value for 7, 11, etc. Because we cannot send the actual value digitally. But, in this case, 36 cannot be represented as a 4-bit value, neither can -36.